requirements.txt
tensorflow-cpu==2.15.0
keras==2.15.0
numpy==1.26.4
scikit-learn
joblib
nltk
vaderSentiment
flask
flask-cors
flask-restx
requests
